Fixed C# program
