--- CSharp/MyApplication/bin/Debug/net9.0/MyApplication.deps.json
@@ -1,23 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v9.0",
-    "signature": ""
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETCoreApp,Version=v9.0": {
-      "MyApplication/1.0.0": {
-        "runtime": {
-          "MyApplication.dll": {}
-        }
-      }
-    }
-  },
-  "libraries": {
-    "MyApplication/1.0.0": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    }
-  }
-}
