+++ CSharp/SchoolApp/bin/Debug/net9.0/SchoolApp.deps.json
@@ -1,0 +1,23 @@
+{
+  "runtimeTarget": {
+    "name": ".NETCoreApp,Version=v9.0",
+    "signature": ""
+  },
+  "compilationOptions": {},
+  "targets": {
+    ".NETCoreApp,Version=v9.0": {
+      "SchoolApp/1.0.0": {
+        "runtime": {
+          "SchoolApp.dll": {}
+        }
+      }
+    }
+  },
+  "libraries": {
+    "SchoolApp/1.0.0": {
+      "type": "project",
+      "serviceable": false,
+      "sha512": ""
+    }
+  }
+}
