+++ CSharp/SchoolApp/bin/Debug/net9.0/SchoolApp.runtimeconfig.json
