+++ CognizantExcersises.sln
@@ -4,7 +4,7 @@ VisualStudioVersion = 17.5.2.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "CSharp", "CSharp", "{B41BF331-FCCB-2ADF-CDB6-767964B34647}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MyApplication", "CSharp\MyApplication\MyApplication.csproj", "{6F735EF0-5C4A-7688-E8C0-5091AE629AB1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SchoolApp", "CSharp\SchoolApp\SchoolApp.csproj", "{6CE3E4D1-96B9-7C78-3A48-C95460EE3421}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -12,18 +12,18 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6F735EF0-5C4A-7688-E8C0-5091AE629AB1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6F735EF0-5C4A-7688-E8C0-5091AE629AB1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6F735EF0-5C4A-7688-E8C0-5091AE629AB1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6F735EF0-5C4A-7688-E8C0-5091AE629AB1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6CE3E4D1-96B9-7C78-3A48-C95460EE3421}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6CE3E4D1-96B9-7C78-3A48-C95460EE3421}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6CE3E4D1-96B9-7C78-3A48-C95460EE3421}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6CE3E4D1-96B9-7C78-3A48-C95460EE3421}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{6F735EF0-5C4A-7688-E8C0-5091AE629AB1} = {B41BF331-FCCB-2ADF-CDB6-767964B34647}
+		{6CE3E4D1-96B9-7C78-3A48-C95460EE3421} = {B41BF331-FCCB-2ADF-CDB6-767964B34647}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {448268EF-4870-427A-9BE3-6C7A640C4333}
+		SolutionGuid = {30E9CE59-D66F-4D4C-B392-5D22BA71A21C}
 	EndGlobalSection
 EndGlobal
